+++ cmd/switch.go
@@ -2,7 +2,7 @@ package cmd
 
 import (
 	"fmt"
-	"github.com/richardjennings/mygit/pkg/mygit"
+	"github.com/richardjennings/mygit/pkg/mygit/git"
 	"github.com/spf13/cobra"
 	"log"
 	"os"
@@ -15,7 +15,7 @@ var switchCmd = &cobra.Command{
 		if err := configure(); err != nil {
 			log.Fatalln(err)
 		}
-		if err := mygit.SwitchBranch(args[0]); err != nil {
+		if err := git.SwitchBranch(args[0]); err != nil {
 			fmt.Println(err)
 			os.Exit(1)
 		}
