+++ pkg/mygit/git/add_test.go
@@ -1,0 +1,1 @@
+package git
