+++ pkg/mygit/git/commit_test.go
@@ -1,0 +1,1 @@
+package git
