+++ pkg/mygit/git/log_test.go
@@ -1,0 +1,2 @@
+package git
+
