+++ pkg/mygit/git/ls-files_test.go
@@ -1,0 +1,1 @@
+package git
