+++ pkg/mygit/git/restore_test.go
@@ -1,0 +1,1 @@
+package git
