+++ pkg/mygit/git/switch_test.go
@@ -1,0 +1,1 @@
+package git
